moved the overlapping functionality to own file, added H check
