+++ doc/nnls.rst
@@ -1,0 +1,8 @@
+Regularized Nonnegative Least Squares
+=====================================
+
+.. code-block:: python
+
+    N = 10
+    
+
