+++ tests/other_tests/test_solve_mpqp.py
@@ -100,13 +100,13 @@ def test_solve_geometric_portfolio(portfolio_problem_analog):
     sol_combi = solve_mpqp(portfolio_problem_analog, mpqp_algorithm.combinatorial)
 
     # they should have the same number of critical regions
-    assert(len(sol_geo) == len(sol_combi))
+    assert (len(sol_geo) == len(sol_combi))
 
     # test the center of each critical region
     for cr in sol_geo.critical_regions:
 
         chev_sol = chebyshev_ball(cr.E, cr.f, deterministic_solver=portfolio_problem_analog.solver.solvers['lp'])
-        center = chev_sol.sol[0].reshape(-1,1)
+        center = chev_sol.sol[0].reshape(-1, 1)
 
         geo_ans = sol_geo.evaluate(center)
         combi_ans = sol_combi.evaluate(center)
